refactor: Move GitHub repository UI components to shared location
Refactors the GitHub repository selection UI components to a common location to reduce duplication and improve maintainability. This consolidates UI components that were previously duplicated across trigger and action node property panels.

- Moved select-repository.tsx and github-repository-block.tsx to shared ui directory
- Updated imports in affected files
- Simplified the component structure to facilitate reuse
