--- internal-packages/workflow-designer-ui/src/editor/properties-panel/action-node-properties-panel/ui/index.ts
@@ -1,2 +1,0 @@
-export * from "./select-repository";
-export * from "./github-repository-block";
