+++ internal-packages/workflow-designer-ui/src/editor/properties-panel/trigger-node-properties-panel/ui/index.ts
@@ -1,3 +1,1 @@
-export * from "./select-repository";
 export * from "./configured-view";
-export * from "./github-repository-block";
