+++ internal-packages/workflow-designer-ui/src/editor/properties-panel/ui/select-repository.tsx
@@ -15,7 +15,7 @@ import {
 	SelectItem,
 	SelectTrigger,
 	SelectValue,
-} from "../../../../ui/select";
+} from "../../../ui/select";
 
 interface SelectRepository {
 	installationId: number;
